Moved unit tests and modified import paths
Most unit tests still fail but it appears that these modifications fix
all of the issues with not being able to find modules.
